% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_summary_function.R
\name{fdr_summary}
\alias{fdr_summary}
\title{Identify enriched events using the p-values}
\usage{
fdr_summary(refsegs_dir, simulation_dir, tumour_type, CNA_type, output_dir)
}
\arguments{
\item{refsegs_dir}{Full path to the directory with the refsegs files}

\item{simulation_dir}{Full path to the directory where the simulations of the CNA events are saved}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{CNA_type}{String that represents the type of CNA we summarise the simulations of (could be LOH, gain or HD)}

\item{output_dir}{Full path to the directory where the files with the enriched regions (and the p-values) will be output}
}
\value{
files with enriched regions, the Bonferroni and FDR-corrected p-values
}
\description{
Identify enriched events using the p-values
}
