% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_enriched_regions.R
\name{identify_enriched_regions}
\alias{identify_enriched_regions}
\title{Simulate random LOH, gains and HD to identify enriched events}
\usage{
identify_enriched_regions(
  annotated_segments_file,
  refsegs_dir,
  gain_output_dir,
  loh_output_dir,
  hd_output_dir,
  tumour_type,
  run
)
}
\arguments{
\item{annotated_segments_file}{Full path to the aggregated subclones file, which has been annotated with the CNA type}

\item{refsegs_dir}{Full path to the directory with the refsegs files}

\item{gain_output_dir}{Full path to the directory where the simulations of the gains will be saved}

\item{loh_output_dir}{Full path to the directory where the simulations of the losses of heterozygosity will be saved}

\item{hd_output_dir}{Full path to the directoey where the simulations of the homozygous deletions will be saved}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{run}{Number of simulation}
}
\value{
files with the random placing of the LOH/gain/HD events
}
\description{
Simulate random LOH, gains and HD to identify enriched events
}
