% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_enriched_regions_for_ordering.R
\name{merge_enriched_regions}
\alias{merge_enriched_regions}
\title{Merge segments overlapping the enriched regions}
\usage{
merge_enriched_regions(
  annotated_segments_file,
  tumour_type,
  enriched_dir,
  hg19_chr_coordinates,
  output_dir
)
}
\arguments{
\item{annotated_segments_file}{Full path to the file with the annotated segments file}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{enriched_dir}{Full path to the directory with the files with the enriched regions}

\item{hg19_chr_coordinates}{Full path to the file with the chromosome coordinates of the hg19 build genome}

\item{output_dir}{Full path to the directory where the merged enriched segments and the plots of the enriched regions are output}
}
\value{
files with merged enriched regions (separate files for LOH, HD and gain) and plots of the segments overlapping the enriched regions
}
\description{
Merge segments overlapping the enriched regions
}
