% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_enriched_regions_for_ordering.R
\name{multipcf_new_breakpoints}
\alias{multipcf_new_breakpoints}
\title{Check if any new breakpoints should be added to the enriched regions}
\usage{
multipcf_new_breakpoints(enriched_dir, tumour_type, output_dir)
}
\arguments{
\item{enriched_dir}{Full path to the directory with the files with the enriched regions}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{output_dir}{Full path to the directory where the enriched segments with the new breakpoints and the plots of start and end positions of the enriched regions are output}
}
\value{
files with merged enriched regions (separate files for LOH, HD and gain) and plots of the segments overlapping the enriched regions
}
\description{
Check if any new breakpoints should be added to the enriched regions
}
