% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_events_across_cohort.R
\name{order_events_across_chort}
\alias{order_events_across_chort}
\title{Process the enriched data into the format required for the ordering step with the Plackett-Luce model}
\usage{
order_events_across_chort(
  annotated_segments_file,
  tumour_type,
  merged_segments_dir,
  output_dir,
  driver_mutations,
  drivers_file = NULL,
  mixture_model,
  model,
  clonal_driver_mutations = NULL
)
}
\arguments{
\item{annotated_segments_file}{Full path to the file with the annotated CNA}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{merged_segments_dir}{Full path to the directory with the LOH, Gain and HD mergedsegs files}

\item{output_dir}{Full path to the directory where the output files and plots are stored}

\item{driver_mutations}{A logical: if TRUE, we include the driver mutations to the model}

\item{drivers_file}{Full path to the file with the drivers mutation information; default NULL}

\item{mixture_model}{A logical: if TRUE, it will run a mixture model, and if FALSE, it will run either the PlackettLuce or the PLMIX implementation of the Plackett Luce}

\item{model}{A string: it could be either PlackettLuce or PLMIX; if it is PlackettLuce, then we don't include the unobserved events, if it is PLMIX we include the unobserved events after the last observed events}

\item{clonal_driver_mutations}{A vector: the list of driver mutations to be included in the model; should be of the format c('cMut_TP53','cMut_ARID1A'), for example; default = NULL}
}
\value{
files with a matrix with true & ordered events, a matrix with the values/ranking of the events, all segments with any CNA, a plot with the overall timing of CNA
}
\description{
Process the enriched data into the format required for the ordering step with the Plackett-Luce model
}
