% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_subclones_for_timing.R
\name{plot_CN_landscape}
\alias{plot_CN_landscape}
\title{Plot the CNA data across the genome (all/clonal/subclonal aberrations)}
\usage{
plot_CN_landscape(
  annotated_segments_file,
  refsegs_dir,
  tumour_type,
  output_dir
)
}
\arguments{
\item{annotated_segments_file}{Full path to the aggregated subclones file, which has been annotated with the CNA type}

\item{refsegs_dir}{Full path to the directory with the refsegs files}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{output_dir}{Full path to the directory where the landscape plots will be saved}
}
\value{
a pdf file with the overall landscape of all/clonal/subclonal events
}
\description{
Plot the CNA data across the genome (all/clonal/subclonal aberrations)
}
