% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_subclones_for_timing.R
\name{prepare_data_for_landscape}
\alias{prepare_data_for_landscape}
\title{Prepare data for plotting of landscape of CNAs across the whole genome.}
\usage{
prepare_data_for_landscape(annotated_segments_file, tumour_type, output_dir)
}
\arguments{
\item{annotated_segments_file}{Full path to the file with the aggregated subclones information, which has been annotated with the CNA type}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{output_dir}{Full path to the directory with the refsegs files output from this function}
}
\value{
files with information about which positions of the genome are aberrated across the whole cohort. There will be a separate file for all/clonal/subclonal LOH/gain/HD.
}
\description{
Prepare data for plotting of landscape of CNAs across the whole genome.
}
