% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_subclones_for_timing.R
\name{subclone_collation}
\alias{subclone_collation}
\title{Collate the subclones data from Battenberg into a single file and add the ploidy of the sample.}
\usage{
subclone_collation(data_dir, tumour_type, output_dir)
}
\arguments{
\item{data_dir}{Full path to the directory with the subclones files}

\item{tumour_type}{String that represents the type of tumour we work with}

\item{output_dir}{Full path to the directory where the allsegs.txt file output from this function will be stored}
}
\value{
\emph{tumour_type}_allsegs.txt file with information about the samplenames, copy number aberrations (startpos, endpos, copy number state of major/minor allele), ploidy
}
\description{
Collate the subclones data from Battenberg into a single file and add the ploidy of the sample.
}
